// portfolio/sanity.types.ts
export type SanityImageCrop = {
  _type: "sanity.imageCrop";
  top?: number;
  bottom?: number;
  left?: number;
  right?: number;
};

export type SanityImageHotspot = {
  _type: "sanity.imageHotspot";
  x?: number;
  y?: number;
  height?: number;
  width?: number;
};
// ...
export type Slug = {
  _type: "slug";
  current?: string;
  source?: string;
};
// ...
export declare const internalGroqTypeReferenceTo: unique symbol;
// ...
export type PROJECTS_QUERYResult = Array<{
  title: string | null;
  slug: Slug | null;
  tagline: string | null;
  category:
    | "ai-ml"
    | "api-backend"
    | "browser-extension"
    | "cli-tool"
    | "desktop-app"
    | "devops"
    | "game"
    | "mobile-app"
    | "open-source"
    | "other"
    | "web-app"
    | null;
  liveUrl: string | null;
  githubUrl: string | null;
  coverImage: {
    asset?: {
      _ref: string;
      _type: "reference";
      _weak?: boolean;
      [internalGroqTypeReferenceTo]?: "sanity.imageAsset";
    };
    media?: unknown;
    hotspot?: SanityImageHotspot;
    crop?: SanityImageCrop;
    alt?: string;
    _type: "image";
  } | null;
  technologies: Array<{
    name: string | null;
    category:
      | "ai-ml"
      | "backend"
      | "cloud"
      | "database"
      | "design"
      | "devops"
      | "frontend"
      | "mobile"
      | "other"
      | "soft-skills"
      | "testing"
      | "tools"
      | null;
    color: string | null;
  }> | null;
}>;